# dfa: Performance improvement by originally treating calls as unit type calls