Only when a feature has fields that are read, new calls for each context will be created.

Even though this decreases the number of values produces significantly, this improves the DFA performance only slightly, overall performances is about 4% better: 

|Test | new | old |
|-|-|-|
|Building fzweb as a jar | 1:12.29s | 1:15.66s |
|Run tests jvm/c/int | 38m35,346s | 40m14,854s|